% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bdiv}
\alias{bdiv}
\title{Plots beta diversity calculations}
\usage{
bdiv(meta.frame, data.types, beta.frame)
}
\arguments{
\item{meta.frame}{A data frame containing all data. Two columns must be titled 'adiv' and
'nseqs' and must contain the alpha diversity values and number of sequences respectively. All
other information be metadata or categorical/continuous variables that describe the samples.}

\item{data.types}{A list containing text indicators of each meta.frame column data type
in order. Input text for each list element are 'cat', 'cont', and 'meta'.}

\item{beta.frame}{A data frame containing all beta diversity calculations for each sample
combination. Data frame should be formatted as "sample1", "sample2", "bdiv", "cmin", "cmax".
The cmin and cmax values represent the confidence range for the beta diversity value.}
}
\value{
No output is returned. Graph file generated: "AphaDiersity.pdf"
}
\description{
This function generates files, plotting between-sample diversity,
i.e. beta diversity, that plots the thetayc values of all catergorical
variable combinations.
}
\examples{
meta.frame <- read.table("IECG57L01.graphics_data.txt", skip=1, header=TRUE, stringsAsFactors=FALSE)
data.types <- strsplit(readLines("IECG57L01.graphics_data.txt")[1], '\\t')[[1]]
beta.frame <- read.table("IECG57L01.beta_data.out", header=TRUE, stringsAsFactors=FALSE)
bdiv(meta.frame, data.types, beta.frame)
}
\keyword{keywords}

