% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numseq}
\alias{numseq}
\title{Plots number of sequences}
\usage{
numseq(meta.frame, data.types)
}
\arguments{
\item{meta.frame}{A data frame containing all data. Two columns must be titled 'adiv' and
'nseqs' and must contain the alpha diversity values and number of sequences respectively. All
other information be metadata or categorical/continuous variables that describe the samples.}

\item{data.types}{A list containing text indicators of each meta.frame column data type
in order. Input text for each list element are 'cat', 'cont', and 'meta'.}
}
\value{
No output is returned. Graph file generated: "NumSequences.pdf"
}
\description{
This function generates files, plotting each independent variable,
whether it be categorical or continuous, versus the calculated
number of sequences.
}
\examples{
meta.frame <- read.table("IECG57L01.graphics_data.txt", skip=1, header=TRUE, stringsAsFactors=FALSE)
data.types <- strsplit(readLines("IECG57L01.graphics_data.txt")[1], '\\t')[[1]]
numseq(meta.frame, data.types)
}
\keyword{keywords}

